razorpay, profile , emailTemplate

+++ Server/controllers/Section.js
@@ -1,5 +1,6 @@
 const Section = require("../models/Section");
 const Course = require("../models/Courses");
+const SubSection = require("../models/SubSection");
 
 exports.createSection = async (req, res) => {
 	try {
@@ -87,25 +88,47 @@ exports.updateSection = async (req, res) => {
 
 // DELETE a section
 exports.deleteSection = async (req, res) => {
-	try {
-		const { sectionId,courseId } = req.body;
+  try {
+    const { sectionId, courseId } = req.body;
+    await Course.findByIdAndUpdate(courseId, {
+      $pull: {
+        courseContent: sectionId,
+      },
+    })
+    const section = await Section.findById(sectionId);
+    console.log(sectionId, courseId);
+    if (!section) {
+      return res.status(404).json({
+        success: false,
+        message: "Section not found",
+      })
+    }
+    // Delete the associated subsections
+    await SubSection.deleteMany({ _id: { $in: section.subSection } })
 
-		await Section.findByIdAndDelete(sectionId);
+    await Section.findByIdAndDelete(sectionId)
 
-		const updatedCourse = await Course.findById(courseId).populate({
-             path: "courseContent", populate: { path: "subSection" } 
-            }).exec();
+    // find the updated course and return it
+    const course = await Course.findById(courseId)
+      .populate({
+        path: "courseContent",
+        populate: {
+          path: "subSection",
+        },
+      })
+      .exec()
 
-		return res.status(200).json({
-			success: true,
-			message: "Section deleted",
-			updatedCourse,
-		});
-	} catch (error) {
-		console.error("Error deleting section:", error);
-		return res.status(500).json({
-			success: false,
-			message: "Internal server error",
-		});
-	}
-};
+    res.status(200).json({
+      success: true,
+      message: "Section deleted",
+      data: course,
+    })
+  } catch (error) {
+    console.error("Error deleting section:", error)
+    res.status(500).json({
+      success: false,
+      message: "Internal server error",
+      error: error.message,
+    })
+  }
+}
